Pension 104893 Reorganize module folders
