+++ modules/pensions/lib/pensions/engine.rb
@@ -18,14 +18,14 @@ module Pensions
 
     initializer 'pensions.military_information' do |app|
       app.config.to_prepare do
-        require_relative '../pensions/military_information'
+        require 'pensions/military_information'
       end
     end
 
     initializer 'pensions.pdf_fill.register_form' do |app|
       app.config.to_prepare do
         require 'pdf_fill/filler'
-        require_relative '../pdf_fill/va21p527ez'
+        require 'pensions/pdf_fill/va21p527ez'
 
         # Register our Pension Pdf Fill form
         ::PdfFill::Filler.register_form(Pensions::FORM_ID, Pensions::PdfFill::Va21p527ez)
@@ -35,7 +35,7 @@ module Pensions
     initializer 'pensions.benefits_intake.register_handler' do |app|
       app.config.to_prepare do
         require 'lighthouse/benefits_intake/sidekiq/submission_status_job'
-        require_relative '../benefits_intake/submission_handler'
+        require 'pensions/benefits_intake/submission_handler'
 
         # Register our Pension Benefits Intake Submission Handler
         ::BenefitsIntake::SubmissionStatusJob.register_handler(Pensions::FORM_ID,
